% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkey_cook.R
\name{cook_turkey}
\alias{cook_turkey}
\title{Return timer in minutes to cook time of a turkey of a given size}
\usage{
cook_turkey(pounds)
}
\arguments{
\item{pounds}{Turkey weight in pounds.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Return timer in minutes to cook time of a turkey of a given size
}
\examples{
cook_turkey(10)
}
